If chk file exists and log does not for opt, then would crash for gaussian. This has been fixed. (for manual file changes)

+++ PoltypeModules/optimization.py
@@ -429,7 +429,7 @@ def GeometryOptimization(poltype,mol,checkbonds=True,modred=True,restraints=None
             jobtolog={cmdstr:os.getcwd()+r'/'+poltype.logfname}
             scratchdir=poltype.scrtmpdirgau
             jobtologlistfilepathprefix=os.getcwd()+r'/'+'optimization_jobtolog_'+poltype.molecprefix 
-            if os.path.isfile(poltype.chkoptfname):
+            if os.path.isfile(poltype.chkoptfname) and os.path.isfile(poltype.logoptfname):
                 os.remove(poltype.logoptfname) # if chk point exists just remove logfile, there could be error in it and we dont want WaitForTermination to catch error before job is resubmitted by daemon 
             if poltype.externalapi==None:
                 finishedjobs,errorjobs=poltype.CallJobsSeriallyLocalHost(jobtooutputlog,True) # have to skip errors because setting optmaxcycle to low number in gaussian causes it to crash
